services/inference/requirements.txt
fastapi
uvicorn[standard]
redis
psycopg2-binary
sqlalchemy
mlflow
lightgbm
python-dotenv
